backend/requirements.txt
django==5.2.6
djangorestframework~=3.15
djangorestframework-simplejwt~=5.3
django-cors-headers~=4.4
# Pillow~=10.4
cryptography~=43.0
bcrypt~=4.2
argon2-cffi~=23.1
gunicorn~=22.0
psycopg2-binary~=2.9
whitenoise~=6.6
stripe
mercadopago
google-auth
requests
lxml
signxml

#coreapi for documentation - Dev: Juan Obando.
coreapi~=2.3
pkg-resources # from setuptools v81.0.0 - exec: pip install setuptools==81.0.0
legacy-cgi
